Redirect /posts to /elections
