+++ ynr/apps/candidates/templates/candidates/areas.html
@@ -15,6 +15,6 @@
 {% block content %}
 
 <p>The election area page has been deprecated. You should update your
-  bookmarks, maybe to <a href="{% url "posts" %}">one of the post pages</a>?</p>
+  bookmarks, maybe to <a href="{% url "election_list_view" %}">one of the post pages</a>?</p>
 
 {% endblock %}
