+++ ynr/apps/candidates/views/posts.py
@@ -1,14 +1,3 @@
-from django.views.generic import TemplateView
-
-from elections.models import Election
 
 
-class PostListView(TemplateView):
-    template_name = "candidates/posts.html"
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["elections_and_posts"] = Election.group_and_order_elections(
-            include_postextraelections=True, include_noncurrent=False
-        )
-        return context
