+++ ynr/apps/elections/templates/elections/election_list.html
