+++ ynr/apps/elections/uk/templates/base.html
@@ -32,7 +32,7 @@
 
 {% block menu_bar_links %}
   <li class="nav-links__item"><a href="{% url 'help-about' %}">{% trans "About" %}</a></li>
-  <li class="nav-links__item"><a href="{% url 'posts' %}">{% trans "Elections" %}</a></li>
+  <li class="nav-links__item"><a href="{% url 'election_list_view' %}">{% trans "Elections" %}</a></li>
   <li class="nav-links__item"><a href="{% url 'reports_home' %}">{% trans "Numbers" %}</a></li>
   <li class="nav-links__item"><a href="{% url 'help-api' %}">{% trans "Get the data" %}</a></li>
   <li class="nav-links__item"><a href="{% url 'results-home' %}" style="color:#FFFF8E">{% trans "Beta: Local election results" %}</a></li>
