+++ ynr/apps/elections/urls.py
@@ -4,8 +4,13 @@ from elections import views
 
 urlpatterns = [
     url(
+        "elections/$",
+        views.ElectionListView.as_view(),
+        name="election_list_view",
+    ),
+    url(
         "elections/(?P<election>[^/]+)/$",
         views.ElectionView.as_view(),
         name="election_view",
-    )
+    ),
 ]
