+++ ynr/apps/elections/views.py
@@ -1,5 +1,5 @@
 from django.db.models import Prefetch
-from django.views.generic import DetailView
+from django.views.generic import DetailView, TemplateView
 from candidates.models import PostExtraElection
 from elections.models import Election
 from popolo.models import Membership
@@ -27,4 +27,15 @@ class ElectionView(DetailView):
             )
         )
 
+        return context
+
+
+class ElectionListView(TemplateView):
+    template_name = "elections/election_list.html"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["elections_and_posts"] = Election.group_and_order_elections(
+            include_postextraelections=True, include_noncurrent=False
+        )
         return context
