+++ ynr/apps/ynr_refactoring/urls.py
@@ -8,5 +8,8 @@ urlpatterns = [
         r"^election/(?P<election>[^/]+)/constituencies$",
         views.RedirectConstituencyListView.as_view(),
         name="constituencies_redirect",
-    )
+    ),
+    url(
+        r"^posts$", views.RedirectPostsListView.as_view(), name="posts_redirect"
+    ),
 ]
